+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed: When slot belongs to single-day entry, no date is saved
 
 * Changed: DB Schema is not synced but only created when none is available
+* Fixed: Close Settingsmenu on clicking item
 
 ## [v0.10.0] - 2018-12-04